ArduinoECG/requirements.txt
pySerial
python-socketio[client]
#python-socketio[asyncio_client]
#asyncio